\name{simul.wsvm}
\alias{simul.wsvm}
\title{ Generating simulation data for weighted svm}
\description{
    This function is to Generate simulation data for weighted svm.
}
\usage{
    simul.wsvm(set.seeds = 123, obs.num = 1000, mu = c(0,0), cov.mat = 0.2*diag(2), X = NULL, Y = NULL)
}
\seealso{\code{\link{wsvm}}}
\keyword{ internal}
