\name{wSVM-package}
\alias{wSVM-package}
\alias{wSVM}
\docType{package}
\title{ Weigthed SVM with boosting algorithm for improving accuracy}
\description{
 We propose weighted SVM methods with penalization form. By adding weights to loss term, we can build up weighted SVM easily and examine classification algorithm properties under weighted SVM. Through comparing each of test error rates, we conclude that our Weighted SVM with boosting has predominant properties than the standard SVM have, as a whole.
}
\details{
    \tabular{ll}{
        Package: \tab wSVM\cr
        Type: \tab Package\cr
        Version: \tab 0.1-7\cr
        Date: \tab 2010-10-03\cr
        License: \tab GPL-2\cr
        LazyLoad: \tab yes\cr
    }
}
\author{
    SungHwan Kim \email{swiss747@korea.ac.kr}\cr 
    Soo-heang Eo \email{hanansh@korera.ac.kr}\cr
}
\seealso{
    \code{\link{wsvm}}, \code{\link{wsvm.predict}}, \code{\link{wsvm.boost}}
}
\keyword{package}
